DEVEXP-423: Show "Evaluate Current XQuery Module" errors in a popup
Updated and standardized the way the XQY debugger reports errors.
This also required some refactoring to the MlxprsError object. It was moved to a new file that is independent of the 'vscode' library so that it could be more easily included in the Debug Client code.
