+++ client/errorReporter.ts
@@ -16,47 +16,10 @@
 
 import { window } from 'vscode';
 
-export interface MlxprsError {
-    popupMessage: string;
-    reportedMessage: string;
-    stack: string;
-    code?: string
-}
+import { MlxprsError } from './mlxprsErrorBuilder';
 
 export class ErrorReporter {
     static mlxprsOutput = window.createOutputChannel('mlxprs');
-
-    static buildMlxprsErrorFromError(error: Error, popupMessageBase: string): MlxprsError {
-        const mlxprsError: MlxprsError = {
-            reportedMessage: 'Unable to determine the error message',
-            stack: 'Unable to find the stack trace in the error object',
-            popupMessage: `${popupMessageBase}`
-        };
-        if (error['reason']) {
-            mlxprsError.popupMessage = `${popupMessageBase}:${error['reason']}`;
-        }
-        if (error.message) {
-            mlxprsError.reportedMessage = error.message;
-            mlxprsError.popupMessage = `${popupMessageBase}:${error.message}`;
-        }
-        if (error.stack) {
-            mlxprsError.stack = error.stack;
-        }
-        if (error['code']) {
-            mlxprsError.code = error['code'];
-        }
-        if (error['statusCode']) {
-            mlxprsError.code = error['statusCode'];
-        }
-        if (error['body']) {
-            if (error['body']['errorResponse']) {
-                mlxprsError.reportedMessage = error['body']['errorResponse']['message'];
-                mlxprsError.code = error['body']['errorResponse']['messageCode'];
-                mlxprsError.popupMessage = `${popupMessageBase}:${error['body']['errorResponse']['status']}`;
-            }
-        }
-        return mlxprsError;
-    }
 
     static reportError(mlxprsError: MlxprsError) {
         ErrorReporter.mlxprsOutput.appendLine(`Error Message: ${mlxprsError.reportedMessage}`);
